refactor(cache) move cache to its own dir
